pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "media-feed"
version = "1.0.0"
description = "RSS feed generator mainly for Chaos Communication Congresses."
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "defusedxml>=0.7.1",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
    "types-pyyaml",
    "types-requests",
]

[project.scripts]
media-feed = "media_feed.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/media_feed"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
